Create separate feature for unsubscriptions and refactor spec
